refactor: consolidate table renderers into table_output.go and fix outputFormat cleanup in tests
- Move all 19 print*Table functions from helpers.go, chore_week.go, and calendar_week.go into cmd/table_output.go
- Merge printTableOutputExtended into printTableOutput (split as printTableOutputMore to stay under gocyclo limit)
- Remove printChoreWeekTable and printCalendarWeekTable from their domain files; they now live with other renderers
- Add t.Cleanup(func() { outputFormat = "" }) to 8 tests that mutated global outputFormat without restoring it
