+++ cmd/helpers_test.go
@@ -216,6 +216,7 @@ func captureStdout(fn func()) string {
 }
 
 func TestPrintOutputJSONFallback(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputJSON
 	output := captureStdout(func() {
 		printOutput(map[string]string{"key": "value"})
@@ -226,6 +227,7 @@ func TestPrintOutputJSONFallback(t *testing.T) {
 }
 
 func TestPrintOutputChoresTable(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	chores := []lib.Chore{
 		{ID: "1", Title: "Clean room", Status: "pending", DueDate: "2026-04-28", Points: 10, AssigneeID: "cat1"},
@@ -240,6 +242,7 @@ func TestPrintOutputChoresTable(t *testing.T) {
 }
 
 func TestPrintOutputRewardsTable(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	rewards := []lib.Reward{
 		{ID: "r1", Title: "Ice cream", Points: 50, EmojiIcon: "🍦", Redeemed: false},
@@ -254,6 +257,7 @@ func TestPrintOutputRewardsTable(t *testing.T) {
 }
 
 func TestPrintOutputFramesTable(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	frames := []lib.Frame{
 		{ID: "f1", Name: "Kitchen Frame", TimeZone: "America/New_York"},
@@ -268,6 +272,7 @@ func TestPrintOutputFramesTable(t *testing.T) {
 }
 
 func TestPrintOutputCalendarTable(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	events := []lib.CalendarEvent{
 		{ID: "e1", Title: "Soccer practice", StartAt: "2026-04-28T16:00:00Z", AllDay: false},
@@ -282,6 +287,7 @@ func TestPrintOutputCalendarTable(t *testing.T) {
 }
 
 func TestPrintOutputCategoriesTable(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	cats := []lib.Category{
 		{ID: "c1", Name: "Alice", Color: "blue"},
@@ -296,6 +302,7 @@ func TestPrintOutputCategoriesTable(t *testing.T) {
 }
 
 func TestPrintOutputRewardRedeemedFlag(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	rewards := []lib.Reward{
 		{ID: "r1", Title: "Movie night", Points: 100, Redeemed: true},
@@ -311,6 +318,7 @@ func TestPrintOutputRewardRedeemedFlag(t *testing.T) {
 }
 
 func TestPrintOutputUnknownTypeDefaultsToJSON(t *testing.T) {
+	t.Cleanup(func() { outputFormat = "" })
 	outputFormat = outputTable
 	output := captureStdout(func() {
 		printOutput(map[string]int{"count": 5})
